Fix neurondb.dat writing and remove human mtype == all
* neurondb.dat fix:

The separator is ' ' and not ','
The name of the human cell and not the rat one must be used

* human mtype == 'all' removal:

The excitatory mapping human mtype -> rat mtype given by Natali was specifying

all -> some rat mtypes

I took it from granted and implemented this functionality in the code.
But since excitatory cells she gave me have only one mtype (PC), I should
have changed the mapping to:

PC -> some rat mtypes

This way there is no longer a special case for excitatory cells.

Change-Id: If89f5eba671dddb1cb75dd22a065930b7f9d7cb5